+++ CHANGELOG.md
@@ -10,10 +10,12 @@ This project adheres (well, maybe not quite) to [Semantic Versioning](http://sem
 - Added README with initial API definition
 - Added base class 'index.js' with the first raw object
 - Added package.json
+- Can now add margins to a chart via setter
 
 ### Changed
 - Renamed index.js to d3by5-base-graphs
 - Refactored _parseData to enable easier overwriting of parts of the properties
+
 
 [//]: ##############################################
 <!---